move searchsorted to algorithms.py
